libdvr: ADD schedule recording,wait for the recording to finish,PVR video cannot be played [1/1]
PD#SWPL-74616

Problem:
Add a schedule recording, wait for the recording to finish and PVR
video cannot be played.

Solution:
PMT does not exist when starting a recording. Need to update recording
PID when we get PMT the first time.

Verify:
Verified by QA on AH212 and project QA on AM301.

Signed-off-by: Wentao.MA <wentao.ma@amlogic.com>

Change-Id: Ibbfea665cb30d79a994a1e77c83f83003746b552